refactoring navbar and router
