+++ documentation/catart.txt
@@ -1,0 +1,28 @@
+
+ /^--^\     /^--^\     /^--^\
+ \____/     \____/     \____/
+ /      \   /      \   /      \
+ |        | |        | |        |
+ \__  __/   \__  __/   \__  __/
+ |^|^|^|^|^|^|^|^|^|^|^|^\ \^|^|^|^/ /^|^|^|^|^\ \^|^|^|^|^|^|^|^|^|^|^|^|
+ | | | | | | | | | | | | |\ \| | |/ /| | | | | | \ \ | | | | | | | | | | |
+ ########################/ /######\ \###########/ /#######################
+ | | | | | | | | | | | | \/| | | | \/| | | | | |\/ | | | | | | | | | | | |
+ |_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|_|
+
+Art by Blazej Kozlowski
+       _                        
+       \`*-.                    
+        )  _`-.                 
+       .  : `. .                
+       : _   '  \\               
+       ; *` _.   `*-._          
+       `-.-'          `-.       
+         ;       `       `.     
+         :.       .        \    
+         . \  .   :   .-'   .   
+         '  `+.;  ;  '      :   
+         :  '  |    ;       ;-. 
+         ; '   : :`-:     _.`* ;
+[bug] .*' /  .*' ; .*`- +'  `*' 
+      `*-*   `*-*  `*-*'
