+++ frontend/app/context.go
@@ -4,31 +4,25 @@ import (
 	"godl/backend/events"
 	"godl/backend/model"
 	"godl/frontend/components"
-	"godl/frontend/router"
-	"godl/frontend/screens"
 
 	"github.com/rivo/tview"
 )
 
 type AppContext struct {
 	app    *tview.Application
-	router router.Router
 	navBar *components.NavBarComponent
 }
 
 func (ac *AppContext) initApp() {
-	ac.router.Draw("welcome")
 }
 
 func (ac *AppContext) addRoutes() {
-	ac.router.AddRoute("welcome", screens.NewWelcomeScreen(ac.app, *baseMediator(ac.app), ac.navBar))
 }
 
 func NewAppContext() *AppContext {
 	a := tview.NewApplication()
 	ac := &AppContext{
 		app:    a,
-		router: *router.NewRouter(),
 		navBar: components.NewNavBarComponent(),
 	}
 	ac.addRoutes()
@@ -52,8 +46,6 @@ func baseMediator(app *tview.Application) *events.Mediator {
 }
 
 func (ac *AppContext) setupNavBar() {
-	ac.navBar.AddTab(*components.NewTabComponent("welcome", ac.router.GetResolution("welcome")))
-	ac.navBar.AddTab(*components.NewTabComponent("settings", func() {}))
 	ac.navBar.SetActive("welcome")
 }
